Various cleanups
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11133 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/laserdisc/OggReader.hh
@@ -79,9 +79,9 @@ private:
 	// ogg state
 	ogg_sync_state sync;
 	ogg_stream_state vorbisStream, theoraStream;
-	long audioSerial;
-	long videoSerial;
-	long skeletonSerial;
+	int audioSerial;
+	int videoSerial;
+	int skeletonSerial;
 	unsigned currentOffset;
 	unsigned totalBytes;
